requirements.txt
alembic==0.9.1
appdirs==1.4.3
click==6.7
coverage==4.3.4
Flask==0.12.1
Flask-Cors==3.0.2
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
funcsigs==1.0.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
mock==2.0.0
nose==1.3.7
packaging==16.8
pbr==2.0.0
psycopg2==2.7.1
py==1.4.33
PyJWT==1.4.2
pyparsing==2.2.0
pytest==3.0.7
python-editor==1.0.3
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.9
Werkzeug==0.12.1
